requirements.txt
absl-py==1.4.0
asgiref==3.7.2
boto3==1.28.47
botocore==1.31.47
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.10.0
Django==4.2.5
filelock==3.12.4
fire==0.5.0
fonttools==4.42.1
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
huggingface-hub==0.17.1
idna==2.10
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.1.0
Pillow==9.5.0
protobuf==4.24.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.6
rsa==4.9
s3transfer==0.6.2
sahi==0.11.14
scipy==1.11.2
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.4
supervision==0.14.0
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.178
urllib3==1.26.16
Werkzeug==2.3.7
yolov5==7.0.12
